refactor IWorkspaceService
