+++ modules/back-end/src/Application/Services/IWorkspaceService.cs
@@ -4,7 +4,5 @@ namespace Application.Services;
 
 public interface IWorkspaceService : IService<Workspace>
 {
-    Task<ICollection<Workspace>> GetByEmailAsync(string email);
-
     Task<bool> HasKeyBeenUsedAsync(Guid workspaceId, string key);
 }
